Fix type of threshold variable in inRange function to correct type. And set stdevlimit to 4 to allow for bigger deviation from the average to be able to find more often a baseline (until now only once one error in 31-05-2014.log)

+++ project/Baseline/src/main.c
@@ -32,7 +32,7 @@ _declspec (dllexport) int32_t findBaseline(int32_t start, int32_t end,
 int32_t calculateBaseline(int32_t start, int32_t end, uint16_t array[],
 	const int32_t size, const uint16_t threshold, int16_t *pBaseline,
 	int16_t *pStdev);
-bool inRange(const int32_t threshold, double value);
+bool inRange(const uint16_t threshold, double value);
 
 
 _declspec (dllexport) int32_t findBaseline(int32_t start, int32_t end,
